requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dash==2.18.2
dash-auth==2.0.0
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.12
decorator==5.1.1
executing==2.1.0
filelock==3.13.1
Flask==3.0.3
fsspec==2024.2.0
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
portalocker==3.1.1
prompt_toolkit==3.0.49
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
retrying==1.3.4
samv2==0.0.4
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.21.0
